ai-service/pyproject.toml
[project]
name = "claexa-ai-service"
version = "1.0.0"
description = "Internal AI microservice of Claexa AI"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "boto3>=1.39.3",
    "google-cloud-storage>=2.12.0",
    "google-genai>=1.24.0",
    "google-generativeai>=0.8.5",
    "matplotlib>=3.10.5",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.3.6",
    "pydantic-settings>=2.10.1",
    "httpx>=0.27.2",
    "grpcio>=1.64.0",
    "grpcio-tools>=1.64.0",
    "grpcio-health-checking>=1.64.0",
    "grpcio-reflection>=1.64.0",
    "protobuf>=5.28.2",
    "pinecone>=7.3.0",
    "logfire>=4.4.0",
    "graphviz>=0.21",
    "mcp>=1.13.1",
    "mcp-run-python>=0.0.21",
    "rdkit>=2025.9.1",
    "music21>=9.9.0",
    "watchdog>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "watchdog>=6.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
